+++ src/collections/index.ts
@@ -1,18 +1,18 @@
 import { Managers } from './access/Managers'
 import { Media } from './resources/Media'
-import { ExternalVideos } from './content/ExternalVideos'
+import { ExternalVideos } from './resources/ExternalVideos'
 import { Narrators } from './resources/Narrators'
 import { Meditations } from './content/Meditations'
 import { Music } from './content/Music'
 import { Pages } from './content/Pages'
 import { Lessons } from './content/Lessons'
-import { LessonUnits } from './tags/LessonUnits'
-import { Frames } from './resources/Frames'
+import { LessonUnits } from './content/LessonUnits'
+import { Frames } from './system/Frames'
 import { Clients } from './access/Clients'
 import { MeditationTags } from './tags/MeditationTags'
 import { MediaTags } from './tags/MediaTags'
 import { MusicTags } from './tags/MusicTags'
-import { FileAttachments } from './resources/FileAttachments'
+import { FileAttachments } from './system/FileAttachments'
 
 // Export all collections as an array
 export const collections = [
